apps/api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.133.1
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg[binary]==3.3.3
pyasn1==0.6.2
pycparser==3.0
pydantic[email]==2.*
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.2.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
